prefer-dot-notation: no finding if fixing causes compile error
Fixes: #707
Fixes: #718
