+++ baselines/packages/mimir/api/src/rules/prefer-dot-notation.d.ts
@@ -1,4 +1,4 @@
-import { AbstractRule } from '@fimbul/ymir';
-export declare class Rule extends AbstractRule {
+import { TypedRule } from '@fimbul/ymir';
+export declare class Rule extends TypedRule {
     apply(): void;
 }
