+++ baselines/packages/mimir/test/prefer-dot-notation/default/test.ts.fix
@@ -31,3 +31,15 @@ for (const key of Object.keys(obj))
 
 type T = string[]['length'];
 type V = string['charCodeAt'];
+
+null!['foo'];
+
+class C {
+    static pub = 1;
+    protected static prot = 2;
+    private static priv = 3;
+}
+
+C.pub;
+C['prot'];
+C['priv'];
