+++ baselines/packages/mimir/test/prefer-dot-notation/no-property-access-from-index-signature/invalid.ts.lint
@@ -1,0 +1,1 @@
+window[];
