+++ packages/mimir/src/rules/prefer-dot-notation.ts
@@ -1,9 +1,10 @@
-import { AbstractRule, Replacement, excludeDeclarationFiles } from '@fimbul/ymir';
-import { isTextualLiteral, isValidPropertyAccess } from 'tsutils';
+import { Replacement, excludeDeclarationFiles, TypedRule } from '@fimbul/ymir';
+import { isCompilerOptionEnabled, isTextualLiteral, isTypeFlagSet, isValidPropertyAccess } from 'tsutils';
 import * as ts from 'typescript';
+import { getRestrictedElementAccessError } from '../restricted-property';
 
 @excludeDeclarationFiles
-export class Rule extends AbstractRule {
+export class Rule extends TypedRule {
     public apply() {
         for (const node of this.context.getFlatAst())
             if (node.kind === ts.SyntaxKind.ElementAccessExpression)
@@ -14,7 +15,7 @@ export class Rule extends AbstractRule {
         if (!isTextualLiteral(node.argumentExpression))
             return;
         const {text} = node.argumentExpression;
-        if (!isValidPropertyAccess(text))
+        if (!isValidPropertyAccess(text) || this.hasCompileErrorIfWrittenAsPropertyAccess(node, text))
             return;
 
         this.addFindingAtNode(
@@ -27,5 +28,18 @@ export class Rule extends AbstractRule {
                 ]
                 : Replacement.replace(node.expression.end, node.end, (node.questionDotToken !== undefined ? '?.' : '.') + text),
         );
+    }
+
+    private hasCompileErrorIfWrittenAsPropertyAccess(node: ts.ElementAccessExpression, name: string): boolean {
+        const type = this.checker.getApparentType(this.checker.getTypeAtLocation(node.expression)).getNonNullableType();
+        if (isTypeFlagSet(type, ts.TypeFlags.Any))
+            return false;
+        const symbol = type.getProperty(name);
+        if (symbol === undefined) {
+            if (type.getStringIndexType() === undefined)
+                return true; // should already be a compile error, don't mess with invalid code
+            return isCompilerOptionEnabled(this.context.compilerOptions, 'noPropertyAccessFromIndexSignature');
+        }
+        return getRestrictedElementAccessError(this.checker, symbol, name, node, type) !== undefined;
     }
 }
