+++ packages/mimir/test/no-restricted-property-access/test.ts
@@ -358,3 +358,5 @@ class MyClass extends Object {
         return super['toString']();
     }
 }
+
+null!['prop'];
