+++ packages/mimir/test/prefer-dot-notation/default.test.json
@@ -1,3 +1,3 @@
 {
-  "files": "*.ts"
+  "project": "tsconfig.json"
 }
