+++ packages/mimir/test/prefer-dot-notation/no-property-access-from-index-signature.test.json
@@ -1,0 +1,3 @@
+{
+  "project": "tsconfig.nopropertyaccessfromindexsignature.json"
+}
