+++ packages/mimir/test/prefer-dot-notation/tsconfig.nopropertyaccessfromindexsignature.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "noPropertyAccessFromIndexSignature": true
+  }
+}
